# fix: Optimize maybe_hide_scam_addresses/3 query