## Motivation
```
Limit  (cost=11568282.92..11568282.93 rows=1 width=232) (actual time=152306.087..152351.767 rows=50 loops=1)
  ->  Sort  (cost=11568282.92..11568282.93 rows=1 width=232) (actual time=152194.533..152240.210 rows=50 loops=1)
        Sort Key: t0.block_number DESC, t0.log_index DESC
        Sort Method: top-N heapsort  Memory: 39kB
        ->  Gather  (cost=1034.08..11568282.91 rows=1 width=232) (actual time=1.057..103726.817 rows=351936550 loops=1)
              Workers Planned: 2
              Workers Launched: 1
              ->  Hash Anti Join  (cost=34.08..11567282.81 rows=1 width=232) (actual time=65.922..91754.472 rows=175968275 loops=2)
                    Hash Cond: (t0.token_contract_address_hash = sabm.address_hash)
                    ->  Parallel Seq Scan on token_transfers t0  (cost=0.00..11017436.35 rows=146616635 width=232) (actual time=65.816..60246.915 rows=175968275 loops=2)
                          Filter: block_consensus
                    ->  Hash  (cost=20.70..20.70 rows=1070 width=32) (actual time=0.009..0.012 rows=0 loops=2)
                          Buckets: 2048  Batches: 1  Memory Usage: 16kB
                          ->  Seq Scan on scam_address_badge_mappings sabm  (cost=0.00..20.70 rows=1070 width=32) (actual time=0.008..0.012 rows=0 loops=2)
Planning Time: 0.114 ms
JIT:
  Functions: 23
  Options: Inlining true, Optimization true, Expressions true, Deforming true
  Timing: Generation 1.397 ms, Inlining 46.351 ms, Optimization 126.046 ms, Emission 70.555 ms, Total 244.350 ms
Execution Time: 152352.718 ms
```
## Changelog
- Change `maybe_hide_scam_addresses/3` query
```
Limit  (cost=23.95..242.89 rows=50 width=232) (actual time=0.022..0.056 rows=50 loops=1)
  ->  Index Scan using "token_transfers_block_number_DESC_log_index_DESC_index" on token_transfers t0  (cost=23.95..773405856.42 rows=176624177 width=232) (actual time=0.022..0.053 rows=50 loops=1)
        Filter: (block_consensus AND (NOT (hashed SubPlan 1)))
        SubPlan 1
          ->  Seq Scan on scam_address_badge_mappings  (cost=0.00..20.70 rows=1070 width=32) (actual time=0.001..0.001 rows=0 loops=1)
Planning Time: 0.081 ms
Execution Time: 0.074 ms
```

## Checklist for your Pull Request (PR)

- [ ] If I added new functionality, I added tests covering it.
- [ ] If I fixed a bug, I added a regression test to prevent the bug from silently reappearing again.
- [ ] I checked whether I should update the docs and did so by submitting a PR to [docs repository](https://github.com/blockscout/docs).
- [ ] If I added/changed/removed ENV var, I submitted a PR to [docs repository](https://github.com/blockscout/docs) to update the list of [env vars](https://github.com/blockscout/docs/blob/master/setup/env-variables/README.md) and I updated the version to `master` in the Version column. If I removed variable, I added it to [Deprecated ENV Variables](https://github.com/blockscout/docs/blob/master/setup/env-variables/deprecated-env-variables/README.md) page. After merging docs PR, changes will be reflected in these [pages](https://docs.blockscout.com/setup/env-variables).
- [ ] If I added new DB indices, I checked, that they are not redundant, with PGHero or other tools.
- [ ] If I added/removed chain type, I modified the Github CI matrix and PR labels accordingly.